refactor(assets): merge the asset use-cases into one asset-operations module
Read and browse are grouped functionality over one domain — the
vaultPatcher precedent (two related operations, one file, one
namespace). assetOperations = { readAssetContent, buildAssetListing };
doc surfaces (AGENTS tree + worked-example mentions, ARCHITECTURE,
wiki.json) follow.
