+++ .devin/wiki.json
@@ -50,7 +50,7 @@
     },
     {
       "title": "Vault Operations",
-      "purpose": "The filesystem I/O layer in vault-operations/ that reads, writes, and transforms vault content. Asset use-cases: asset-reader.ts (per-type read dispatch — image fitting, canvas linearize/raw, strict-UTF-8 text passthrough, structured errors) and asset-listing.ts (extension filter, per-extension counts, capped statted slice) compose the vaultFs primitives readAsset (binary read with .md rejection and stat-first size cap), listAssets (folder-scoped non-md walk), and statAssets (bounded-concurrency stat) behind the vault_read_asset/vault_list_assets tools. Safety primitives: atomic writes (atomicWriteFile temp-then-rename, atomicWriteFileExclusive link-based no-clobber for TOCTOU-safe creates), per-file write locking (file-write-lock.ts — serializing, fail-fast, and multi-file fail-fast modes), path traversal prevention (resolveSafePath resolve + prefix check, toVaultRelativePath normalization), and protected-path enforcement (PROTECTED_PATHS checked after normalization). Covers: vault-filesystem.ts (read/write/list/delete .md files, outline and section reads, protected-path validation); vault-patcher.ts (surgical edits — heading-targeted patch with prepend/append/replace operations, and find-and-replace with optional replace-all); note-mover.ts (move/rename a note and rewrite every vault-wide link that pointed to it, using the link resolution engine from obsidian-markdown/links.ts — preflight-then-commit design with multi-file exclusive locking); memory-store.ts (About Me/ heading-aware read/append/delete with shrink guard, idempotency guard, and ambiguity guard — see Memory Layer page); daily-notes.ts (daily note config reader and path resolver — see Memory Layer page). All functions follow the (params, logger) two-argument pattern.",
+      "purpose": "The filesystem I/O layer in vault-operations/ that reads, writes, and transforms vault content. Asset use-case: asset-operations.ts (per-type read dispatch — image fitting, canvas linearize/raw, strict-UTF-8 text passthrough, structured errors — plus browsing: extension filter, per-extension counts, capped statted slice) composes the vaultFs primitives readAsset (binary read with .md rejection and stat-first size cap), listAssets (folder-scoped non-md walk), and statAssets (bounded-concurrency stat) behind the vault_read_asset/vault_list_assets tools. Safety primitives: atomic writes (atomicWriteFile temp-then-rename, atomicWriteFileExclusive link-based no-clobber for TOCTOU-safe creates), per-file write locking (file-write-lock.ts — serializing, fail-fast, and multi-file fail-fast modes), path traversal prevention (resolveSafePath resolve + prefix check, toVaultRelativePath normalization), and protected-path enforcement (PROTECTED_PATHS checked after normalization). Covers: vault-filesystem.ts (read/write/list/delete .md files, outline and section reads, protected-path validation); vault-patcher.ts (surgical edits — heading-targeted patch with prepend/append/replace operations, and find-and-replace with optional replace-all); note-mover.ts (move/rename a note and rewrite every vault-wide link that pointed to it, using the link resolution engine from obsidian-markdown/links.ts — preflight-then-commit design with multi-file exclusive locking); memory-store.ts (About Me/ heading-aware read/append/delete with shrink guard, idempotency guard, and ambiguity guard — see Memory Layer page); daily-notes.ts (daily note config reader and path resolver — see Memory Layer page). All functions follow the (params, logger) two-argument pattern.",
       "parent": "Architecture"
     },
     {
