+++ src/vault-mcp/mcp-core/tools/asset-tools.ts
@@ -1,8 +1,7 @@
 /** Asset tool registration — reading and discovering non-markdown vault files. */
 
 import { z } from "zod"
-import { assetReader } from "../../vault-operations/asset-reader.js"
-import { assetListing } from "../../vault-operations/asset-listing.js"
+import { assetOperations } from "../../vault-operations/asset-operations.js"
 import type { FittedImage } from "../../../utils/fit-image-to-byte-budget.js"
 import type { ToolRegistrationContext } from "./tool-helpers.js"
 import { safeHandler, safeHandlerContent } from "./tool-helpers.js"
@@ -96,7 +95,7 @@ Search coverage: vault_search indexes markdown notes; find assets by browsing (v
       return safeHandlerContent(
         reqLogger,
         () =>
-          assetReader.readAssetContent(
+          assetOperations.readAssetContent(
             {
               vaultPath,
               path,
@@ -197,7 +196,7 @@ Returns: JSON with assets (array of { path, extension, bytes }, sorted by path),
       return safeHandler(
         reqLogger,
         async () => {
-          const listing = await assetListing.buildAssetListing(
+          const listing = await assetOperations.buildAssetListing(
             { vaultPath, folder, extensions, limit: limit ?? 50 },
             reqLogger,
           )
